Move column-level functionality to own module
